% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{two-dimensional numerical interpolation}
\usage{
interp(z, coords, predicts, method = "bilinear", polyorder = 3, na.rm = TRUE)
}
\arguments{
\item{z}{The attribute vector}

\item{coords}{The coordinate matrix}

\item{predicts}{Matrix of locations to predict}

\item{method}{(optional) Interpolation method}

\item{polyorder}{(optional) Degree of the polynomial}

\item{na.rm}{(optional) Whether \code{NA} values should be stripped}
}
\value{
A numeric vector
}
\description{
two-dimensional numerical interpolation
}
\examples{
data("ndvi")
interp(ndvi$NDVI,ndvi[,c("X","Y")],
       matrix(c(2745,2455),ncol = 2))

}
