% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi}
\alias{ndvi}
\title{An NDVI grid dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 121 rows and 3 columns.
}
\usage{
ndvi
}
\description{
An NDVI grid dataset
}
\keyword{datasets}
